requirements.txt
numpy
pandas
scikit-learn
mlflow
dvc
prefect
fastapi
uvicorn[standard]
evidently
pyyaml
joblib
ipykernel
